Refactor `ClientChatStore` into separate stores for modularity and improve maintainability.
